+++ projects/igniteui-angular/src/lib/services/overlay/position/connected-positioning-strategy.ts
@@ -6,7 +6,8 @@ import {
   PositionSettings,
   Point,
   Size,
-  VerticalAlignment
+  VerticalAlignment,
+  cloneObject
 } from './../utilities';
 import { scaleInVerTop, scaleOutVerTop } from '../../../animations/main';
 
@@ -33,6 +34,10 @@ export class ConnectedPositioningStrategy implements IPositionStrategy {
   position(contentElement: HTMLElement, size: Size, document?: Document, initialCall?: boolean): void {
     const startPoint = getPointFromPositionsSettings(this.settings, contentElement.parentElement);
     this.setStyle(contentElement, startPoint, this.settings);
+  }
+
+  clone(): IPositionStrategy {
+    return cloneObject(this);
   }
 
   protected setStyle(contentElement: HTMLElement, startPont: Point, settings: PositionSettings) {