+++ projects/igniteui-angular/src/lib/services/overlay/position/global-position-strategy.ts
@@ -1,5 +1,5 @@
 import { IPositionStrategy } from './IPositionStrategy';
-import { PositionSettings, HorizontalAlignment, VerticalAlignment, Size } from './../utilities';
+import { PositionSettings, HorizontalAlignment, VerticalAlignment, Size, cloneObject } from './../utilities';
 import { fadeIn, fadeOut } from '../../../animations/main';
 
 export class GlobalPositionStrategy implements IPositionStrategy {
@@ -49,6 +49,10 @@ export class GlobalPositionStrategy implements IPositionStrategy {
             default:
                 break;
         }
+    }
+
+    clone(): IPositionStrategy {
+        return cloneObject(this);
     }
 }